# sensor-vista/src/main/java/Actia/sensor_vista/espGroup/EspGroupResponse.java
package Actia.sensor_vista.espGroup;

import lombok.*;

@Getter
@Setter
@AllArgsConstructor
@NoArgsConstructor
@Builder
public class EspGroupResponse {
    private String channelType;
    private String instance;

}